[fix] friends -> friend

+++ friends_service/friend_app/views.py
@@ -13,16 +13,14 @@ from rest_framework.status import (
 )
 from rest_framework.views import APIView
 
-from .models import Friends
+from .models import Friend
 from .serializers import FriendsSerializer, UserIdValidator
 
 
 class FriendListView(APIView):
     def get(self, request):
         user_id = 1
-        friends = Friends.objects.filter(
-            Q(from_user_id=user_id) | Q(to_user_id=user_id)
-        )
+        friends = Friend.objects.filter(Q(from_user_id=user_id) | Q(to_user_id=user_id))
         serializer = FriendsSerializer(friends, many=True)
         friends_data = []
         for friend in serializer.data:
@@ -43,16 +41,16 @@ class FriendRequestView(APIView):
         friend_status = friend.status if friend else None
         rev_friend_status = rev_friend.status if rev_friend else None
 
-        if friend_status == Friends.STATUS_PENDING:
+        if friend_status == Friend.STATUS_PENDING:
             return Response(
                 {"error": "Friend request already sent."}, status=HTTP_409_CONFLICT
             )
         if (
-            friend_status == Friends.STATUS_APPROVED
-            or rev_friend_status == Friends.STATUS_APPROVED
+            friend_status == Friend.STATUS_APPROVED
+            or rev_friend_status == Friend.STATUS_APPROVED
         ):  # すでにfriend
             return Response({"error": "Already friend."}, status=HTTP_400_BAD_REQUEST)
-        if rev_friend_status == Friends.STATUS_PENDING:
+        if rev_friend_status == Friend.STATUS_PENDING:
             return Response(
                 {"error": "Friend requests have already been received."},
                 status=HTTP_409_CONFLICT,
@@ -66,13 +64,13 @@ class FriendRequestView(APIView):
 
         to_user_id = int(user_id_validator.validated_data["user_id"])
         from_user_id = 1
-        friend = Friends.objects.filter(
+        friend = Friend.objects.filter(
             from_user_id=from_user_id, to_user_id=to_user_id
         ).first()
         """
         rev_friend:from_user_idとto_user_idを逆にしたもの
         """
-        rev_friend = Friends.objects.filter(
+        rev_friend = Friend.objects.filter(
             from_user_id=to_user_id, to_user_id=from_user_id
         ).first()
 
@@ -84,10 +82,10 @@ class FriendRequestView(APIView):
                 status=HTTP_400_BAD_REQUEST,
             )
 
-        Friends.objects.create(
+        Friend.objects.create(
             from_user_id=from_user_id,
             to_user_id=to_user_id,
-            status=Friends.STATUS_PENDING,
+            status=Friend.STATUS_PENDING,
         )
         return Response(
             {"message": "Friend request sent successfully."}, status=HTTP_201_CREATED
@@ -104,7 +102,7 @@ class FriendRequestView(APIView):
                 {"error": "You cannot send a request to yourself."},
                 status=HTTP_400_BAD_REQUEST,
             )
-        friend = Friends.objects.filter(
+        friend = Friend.objects.filter(
             from_user_id=from_user_id, to_user_id=to_user_id
         ).first()
         if not friend:
@@ -112,13 +110,11 @@ class FriendRequestView(APIView):
                 {"error": "No friend request exists from the specified user."},
                 status=HTTP_404_NOT_FOUND,
             )
-        if friend.status != Friends.STATUS_PENDING:
+        if friend.status != Friend.STATUS_PENDING:
             return Response(
                 {"error": "Already friend."}, status=HTTP_400_BAD_REQUEST
             )  # すでにfriend
-        Friends.objects.filter(
-            from_user_id=from_user_id, to_user_id=to_user_id
-        ).delete()
+        Friend.objects.filter(from_user_id=from_user_id, to_user_id=to_user_id).delete()
         return Response(status=HTTP_204_NO_CONTENT)
 
     def patch(self, _, user_id):
@@ -134,18 +130,18 @@ class FriendRequestView(APIView):
                 {"error": "You cannot send a request to yourself."},
                 status=HTTP_400_BAD_REQUEST,
             )
-        friend = Friends.objects.filter(
+        friend = Friend.objects.filter(
             from_user_id=from_user_id, to_user_id=to_user_id
         ).first()
         if not friend:
             return Response(
                 {"error": "Friend request not found."}, status=HTTP_404_NOT_FOUND
             )
-        if friend.status == Friends.STATUS_APPROVED:
+        if friend.status == Friend.STATUS_APPROVED:
             return Response(
                 {"error": "Friend request already approved."}, status=HTTP_409_CONFLICT
             )
-        friend.status = Friends.STATUS_APPROVED
+        friend.status = Friend.STATUS_APPROVED
         friend.approved_at = now()
         friend.save()
         return Response({"userId": from_user_id}, status=HTTP_200_OK)
@@ -164,17 +160,17 @@ class FriendView(APIView):
                 status=HTTP_400_BAD_REQUEST,
             )
 
-        friend = Friends.objects.filter(
+        friend = Friend.objects.filter(
             from_user_id=from_user_id, to_user_id=to_user_id
         ).first()
-        rev_friend = Friends.objects.filter(
+        rev_friend = Friend.objects.filter(
             from_user_id=to_user_id, to_user_id=from_user_id
         ).first()
         match (friend is not None, rev_friend is not None):
-            case (True, False) if friend.status == Friends.STATUS_APPROVED:
+            case (True, False) if friend.status == Friend.STATUS_APPROVED:
                 friend.delete()
                 return Response(status=HTTP_204_NO_CONTENT)
-            case (False, True) if rev_friend.status == Friends.STATUS_APPROVED:
+            case (False, True) if rev_friend.status == Friend.STATUS_APPROVED:
                 rev_friend.delete()
                 return Response(status=HTTP_204_NO_CONTENT)
             case _:
